requirements.txt
pydantic
pydantic-ai
httpx
python-dotenv
pdfplumber
streamlit
